+++ Packages/com.hung.data/Runtime/Base/IDataService.cs
@@ -6,6 +6,10 @@ namespace Hung.Base
     {
         public T GetData<T>(int index = 0) where T : class;
         public T GetSOData<T>() where T : ScriptableObject;
+        public T GetSOData<T>(string id = null) where T : ScriptableObject
+        {
+            return GetSOData<T>();
+        }
         public T GetUnit<T>(int type) where T : class;
         public void Save();
     }